Refactor: Changes according to PR comments
