+++ packages/grafana-data/src/types/index.ts
@@ -10,4 +10,3 @@ export * from './utils';
 export * from './valueMapping';
 export * from './displayValue';
 export * from './graph';
-export * from './text';
